+++ src/core/Field.ts
@@ -1,7 +1,0 @@
-import { JSONValue } from '../types';
-
-export default class Field {
-  getSnapshot(): JSONValue {
-    return {};
-  }
-}