+++ src/fields/Field.ts
@@ -1,0 +1,5 @@
+import { JSONValue } from '../types';
+
+export default abstract class Field {
+  abstract getSnapshot(): JSONValue;
+}