+++ src/fields/Person.ts
@@ -1,35 +1,13 @@
-import Field from '../core/Field';
-import { JSONValue, Maybe } from '../types';
-import { cast } from '../utils/parsers';
+import { JSONValue } from '../types';
+import { Info } from './Info';
 
-export class Person extends Field {
-  #email?: string;
-  #url?: string;
-
+export class Person extends Info {
   name: string;
 
-  constructor({ name, url, email }: { name: string; url?: string; email?: string }) {
-    super();
+  constructor({ email, name, url }: { email?: string; name: string; url?: string }) {
+    super({ email, url });
 
     this.name = name;
-    this.url = url;
-    this.email = email;
-  }
-
-  get email(): Maybe<string> {
-    return this.#email;
-  }
-
-  set email(value: Maybe<string>) {
-    this.#email = typeof value === 'string' ? cast.toEmail(value) : value;
-  }
-
-  get url(): Maybe<string> {
-    return this.#url;
-  }
-
-  set url(value: Maybe<string>) {
-    this.#url = typeof value === 'string' ? cast.toUrl(value) : value;
   }
 
   getSnapshot(): JSONValue {