+++ CHANGELOG.md
@@ -1,23 +1,54 @@
 # Important Changes
 
-## License
+## Engines
 
-Source code now under `MIT` license.
+- Bumped **node** from `14.0.0` to `14.13.1`
 
-# :tada: Features
+## Dependencies
 
-- Add Package class for `package.json` parsing [`9bd960f`](https://github.com/keindev/package-json-helper/commit/9bd960fc18e4f25d0a1103ee8403e640e0edfa0f)
-- Add package write replacer [`8a80b5e`](https://github.com/keindev/package-json-helper/commit/8a80b5eaf7de753b28faefdeb5b9ab65b03f4b63)
+<details>
+<summary>Dependencies</summary>
+
+- Added **[execa](https://www.npmjs.com/package/execa/v/6.0.0)** with `^6.0.0`
+
+</details>
+
+<details>
+<summary>Dev Dependencies</summary>
+
+- Bumped **[@tagproject/ts-package-shared-config](https://www.npmjs.com/package/@tagproject/ts-package-shared-config/v/6.0.1)** from `4.0.2` to `6.0.1`
+- Bumped **[@types/jest](https://www.npmjs.com/package/@types/jest/v/27.0.3)** from `27.0.1` to `27.0.3`
+- Bumped **[@types/node](https://www.npmjs.com/package/@types/node/v/16.11.11)** from `16.7.2` to `16.11.11`
+- Bumped **[@types/semver](https://www.npmjs.com/package/@types/semver/v/7.3.9)** from `7.3.8` to `7.3.9`
+- Bumped **[@typescript-eslint/eslint-plugin](https://www.npmjs.com/package/@typescript-eslint/eslint-plugin/v/5.5.0)** from `4.29.3` to `5.5.0`
+- Bumped **[@typescript-eslint/parser](https://www.npmjs.com/package/@typescript-eslint/parser/v/5.5.0)** from `4.29.3` to `5.5.0`
+- Bumped **[cspell](https://www.npmjs.com/package/cspell/v/5.13.1)** from `5.8.2` to `5.13.1`
+- Bumped **[eslint](https://www.npmjs.com/package/eslint/v/8.3.0)** from `7.32.0` to `8.3.0`
+- Bumped **[eslint-plugin-import](https://www.npmjs.com/package/eslint-plugin-import/v/2.25.3)** from `2.24.2` to `2.25.3`
+- Bumped **[eslint-plugin-jest](https://www.npmjs.com/package/eslint-plugin-jest/v/25.3.0)** from `24.4.0` to `25.3.0`
+- Bumped **[eslint-plugin-promise](https://www.npmjs.com/package/eslint-plugin-promise/v/5.1.1)** from `5.1.0` to `5.1.1`
+- Bumped **[husky](https://www.npmjs.com/package/husky/v/7.0.4)** from `7.0.2` to `7.0.4`
+- Bumped **[jest](https://www.npmjs.com/package/jest/v/27.4.2)** from `27.0.6` to `27.4.2`
+- Bumped **[prettier](https://www.npmjs.com/package/prettier/v/2.5.0)** from `2.3.2` to `2.5.0`
+- Bumped **[ts-jest](https://www.npmjs.com/package/ts-jest/v/27.0.7)** from `27.0.5` to `27.0.7`
+- Bumped **[typescript](https://www.npmjs.com/package/typescript/v/4.5.2)** from `4.4.2` to `4.5.2`
+
+</details>
 
 # :fire: Improvements
 
-- Add Parser class, structure parsing codebase [`2da38f2`](https://github.com/keindev/package-json-helper/commit/2da38f25153ce26e1b68619737f3fcbc1f4cb891)
-- Rebuild package parser, add fields classes, bump deps [`e272419`](https://github.com/keindev/package-json-helper/commit/e27241905806fe98235ed002da9edb29c8534187)
-- Finish `package.json` structure processing [`187b661`](https://github.com/keindev/package-json-helper/commit/187b66140fcde89b94856af82384c07a38e6e37e)
+- Add util to compare package deps versions [`0926db5`](https://github.com/keindev/package-json-helper/commit/0926db5483f3863f27d5ef12c82ba27dedf537c9)
+- Add install and deps checks methods [`b1d0018`](https://github.com/keindev/package-json-helper/commit/b1d0018628c8b40de621d54bfda4e08722d72219)
 
-# :memo: Internal changes
+# :bug: Bug Fixes
 
-- Add `PackageBase` test [`db30dae`](https://github.com/keindev/package-json-helper/commit/db30daed8640e3a16443e27dddf2679d3ba9df4d)
+- Add `package.json` field values sorting [`62041bd`](https://github.com/keindev/package-json-helper/commit/62041bd43c9c7faf1fb577eadd8816a1ce8bebb7)
+- Fix a Package constructor value type [`063369b`](https://github.com/keindev/package-json-helper/commit/063369b6f36a85cc92c1b7c5e1df70affe2afb5d)
+
+# :wrench: Code Refactoring
+
+- Fix eslint errors, order fields, add docs [`21cc5e1`](https://github.com/keindev/package-json-helper/commit/21cc5e19a05fff2fa5cdbb88174f268a9c71442b)
+- Fix eslint errors [`53e6af3`](https://github.com/keindev/package-json-helper/commit/53e6af390f91fbcbe1c962d6a76337f51ec057c9)
 
 ---